requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
python-decouple==3.6
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
